Moves tarfile to its own library.
Aids in testing and makes separation of responsibilities clearer.
