+++ listener/listener_test.go
@@ -8,7 +8,7 @@ import (
 	"testing"
 	"time"
 
-	"github.com/m-lab/pusher/tarcache"
+	"github.com/m-lab/pusher/tarfile"
 
 	"github.com/m-lab/pusher/listener"
 )
@@ -20,7 +20,7 @@ func TestListenForClose(t *testing.T) {
 		return
 	}
 	defer os.RemoveAll(dir)
-	ldfChan := make(chan tarcache.LocalDataFile)
+	ldfChan := make(chan tarfile.LocalDataFile)
 	l, err := listener.Create(dir, ldfChan)
 	if err != nil {
 		t.Errorf("%v", err)
@@ -45,7 +45,7 @@ func TestListenForMove(t *testing.T) {
 	defer os.RemoveAll(dir)
 	os.Mkdir(dir+"/subdir", 0777)
 	ioutil.WriteFile(dir+"/testfile", []byte("test"), 0777)
-	ldfChan := make(chan tarcache.LocalDataFile)
+	ldfChan := make(chan tarfile.LocalDataFile)
 	l, err := listener.Create(dir+"/subdir", ldfChan)
 	if err != nil {
 		t.Errorf("%v", err)
@@ -73,7 +73,7 @@ func TestListenInSubdir(t *testing.T) {
 	}
 	defer os.RemoveAll(dir)
 	os.Mkdir(dir+"/subdir", 0777)
-	ldfChan := make(chan tarcache.LocalDataFile)
+	ldfChan := make(chan tarfile.LocalDataFile)
 	l, err := listener.Create(dir+"/subdir", ldfChan)
 	if err != nil {
 		t.Errorf("%v", err)
@@ -99,7 +99,7 @@ func TestCreateOnBadDir(t *testing.T) {
 		return
 	}
 	defer os.RemoveAll(dir)
-	ldfChan := make(chan tarcache.LocalDataFile)
+	ldfChan := make(chan tarfile.LocalDataFile)
 	l, err := listener.Create(dir+"/doesnotexist", ldfChan)
 	if l != nil || err == nil {
 		t.Error("Should have had an error")
