+++ listener/listener_whitebox_test.go
@@ -8,7 +8,7 @@ import (
 	"testing"
 	"time"
 
-	"github.com/m-lab/pusher/tarcache"
+	"github.com/m-lab/pusher/tarfile"
 	"github.com/rjeczalik/notify"
 	"golang.org/x/sys/unix"
 )
@@ -46,7 +46,7 @@ func TestBadEvent(t *testing.T) {
 		return
 	}
 	defer os.RemoveAll(dir)
-	ldfChan := make(chan tarcache.LocalDataFile)
+	ldfChan := make(chan tarfile.LocalDataFile)
 	l, err := Create(dir, ldfChan)
 	if err != nil {
 		t.Errorf("%v", err)
